tests/wpt/tests/css/css-grid/grid-lanes/tentative/parsing/grid-lanes-shorthand-valid.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Grid Lanes: parsing grid-lanes with valid values</title>
  <link rel="author" title="Yanling Wang" href="mailto:yanlingwang@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-grid-3">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
  <script src="/css/support/shorthand-testcommon.js"></script>
</head>
<style>
  #div {
    grid-lanes: "test" max-content row !important;
  }
</style>
<body>
  <div id=div></div>
  <div id=testDiv></div>
  <script>
    test(() => {
      assert_equals(getComputedStyle(div).gridLanes, '"test" max-content row normal');
    }, 'grid-lanes followed by !important');

    function test_valid_grid_lanes_value(property, value, serializedValue) {
      if (arguments.length < 3)
        serializedValue = value;
      var stringifiedValue = JSON.stringify(value);
      test(()=>{
        var testDiv = document.getElementById('testDiv');
        testDiv.style[property] = "";
        testDiv.style[property] = value;
        var readValue = getComputedStyle(testDiv).getPropertyValue(property);
        assert_not_equals(readValue, "", "property should be set");
        assert_equals(readValue, serializedValue, "serialization should be canonical");
      }, `grid-lanes: ${value} should be valid.`);
    }

    test_valid_value("grid-lanes", '"a" calc(10px) column-reverse normal');
    test_valid_value("grid-lanes", 'minmax(calc(30% + 40vw), 10px)', 'minmax(calc(30% + 40vw), 10px) column normal');
    test_valid_grid_lanes_value("grid-lanes", 'minmax(10px, 20px) row', 'minmax(10px, 20px) row normal');
    test_valid_grid_lanes_value("grid-lanes", '1px 2px', '1px 2px column normal');
    test_valid_grid_lanes_value("grid-lanes", '"a" 10px reverse', '"a" 10px column reverse');
    test_valid_grid_lanes_value("grid-lanes", '"a b" 10px 20px row normal');
    test_valid_grid_lanes_value("grid-lanes", '"a b c" 10% 20% 30% row-reverse', '"a b c" 10% 20% 30% row-reverse normal');
    test_valid_grid_lanes_value("grid-lanes", 'repeat(5, auto) row reverse');
    test_shorthand_value('grid-lanes', 'none', {
      'grid-template-columns': 'none',
      'grid-template-areas': 'none',
      'masonry-direction': 'column',
      'grid-lanes-fill': 'normal'
    });
    test_shorthand_value('grid-lanes', '10px reverse', {
      'grid-template-columns': '10px',
      'grid-template-areas': 'none',
      'masonry-direction': 'column',
      'grid-lanes-fill': 'reverse'
    });
    test_shorthand_value('grid-lanes', '"b a" 20% 40% column normal', {
      'grid-template-columns': '20% 40%',
      'grid-template-areas': '"b a"',
      'masonry-direction': 'column',
      'grid-lanes-fill': 'normal'
    });
    test_shorthand_value('grid-lanes', '"b b a" 1fr 2fr 3fr row', {
      'grid-template-rows': '1fr 2fr 3fr',
      'grid-template-areas': '"b" "b" "a"',
      'masonry-direction': 'row',
      'grid-lanes-fill': 'normal'
    });
    test_shorthand_value('grid-lanes', 'repeat(2, auto) row-reverse', {
      'grid-template-rows': 'repeat(2, auto)',
      'grid-template-areas': 'none',
      'masonry-direction': 'row-reverse',
      'grid-lanes-fill': 'normal'
    });
  </script>
</body>
</html>
